// packages/muya/src/state/types.ts
export interface ITableRowState {
    name: 'table.row';
    children: ITableCellState[];
}

export interface ITableCellMeta {
    align: string; // 'none' | 'left' | 'center' | 'right';
}

export interface ITableCellState {
    name: 'table.cell';
    meta: ITableCellMeta;
    text: string;
}

export interface ITableState {
    name: 'table';
    children: ITableRowState[];
}